renaming module and adding initial longdouble backend support
