+++ quaddtype/meson.build
@@ -1,12 +1,12 @@
-project('quaddtype', 'c', 'cpp', default_options : ['cpp_std=c++17', 'b_pie=true'])
+project('numpy_quaddtype', 'c', 'cpp', default_options : ['cpp_std=c++17', 'b_pie=true'])
 
 py_mod = import('python')
 py = py_mod.find_installation()
 
 c = meson.get_compiler('c')
 
-sleef_dep = c.find_library('sleef')
-sleefquad_dep = c.find_library('sleefquad')
+sleef_dep = c.find_library('sleef', dirs:['/usr/local/lib'])
+sleefquad_dep = c.find_library('sleefquad', dirs:['/usr/local/lib'])
 
 incdir_numpy = run_command(py,
   [
@@ -19,30 +19,30 @@ incdir_numpy = run_command(py,
 includes = include_directories(
     [
         incdir_numpy,
-        'quaddtype/src',
+        'numpy_quaddtype/src',
     ]
 )
 
 srcs = [
-    'quaddtype/src/casts.h',
-    'quaddtype/src/casts.cpp',
-    'quaddtype/src/scalar.h',
-    'quaddtype/src/scalar.c',
-    'quaddtype/src/dtype.h',
-    'quaddtype/src/dtype.c',
-    'quaddtype/src/quaddtype_main.c',
-    'quaddtype/src/scalar_ops.h',
-    'quaddtype/src/scalar_ops.cpp',
-    'quaddtype/src/ops.hpp',
-    'quaddtype/src/umath.h',
-    'quaddtype/src/umath.cpp'
+    'numpy_quaddtype/src/casts.h',
+    'numpy_quaddtype/src/casts.cpp',
+    'numpy_quaddtype/src/scalar.h',
+    'numpy_quaddtype/src/scalar.c',
+    'numpy_quaddtype/src/dtype.h',
+    'numpy_quaddtype/src/dtype.c',
+    'numpy_quaddtype/src/quaddtype_main.c',
+    # 'numpy_quaddtype/src/scalar_ops.h',
+    # 'numpy_quaddtype/src/scalar_ops.cpp',
+    # 'numpy_quaddtype/src/ops.hpp',
+    # 'numpy_quaddtype/src/umath.h',
+    # 'numpy_quaddtype/src/umath.cpp'
 ]
 
 py.install_sources(
     [
-        'quaddtype/__init__.py',
+        'numpy_quaddtype/__init__.py',
     ],
-    subdir: 'quaddtype',
+    subdir: 'numpy_quaddtype',
     pure: false
 )
 
@@ -51,6 +51,6 @@ srcs,
 c_args: ['-g', '-O0', '-lsleef', '-lsleefquad'],
 dependencies: [sleef_dep, sleefquad_dep],
 install: true,
-subdir: 'quaddtype',
+subdir: 'numpy_quaddtype',
 include_directories: includes
 )
