+++ quaddtype/numpy_quaddtype/__init__.py
@@ -1,0 +1,23 @@
+from ._quaddtype_main import (
+    QuadPrecision,
+    QuadPrecDType
+)
+
+__all__ = ['QuadPrecision', 'QuadPrecDType', 'SleefQuadPrecision', 'LongDoubleQuadPrecision',
+           'SleefQuadPrecDType', 'LongDoubleQuadPrecDType']
+
+
+def SleefQuadPrecision(value):
+    return QuadPrecision(value, backend='sleef')
+
+
+def LongDoubleQuadPrecision(value):
+    return QuadPrecision(value, backend='longdouble')
+
+
+def SleefQuadPrecDType():
+    return QuadPrecDType(backend='sleef')
+
+
+def LongDoubleQuadPrecDType():
+    return QuadPrecDType(backend='longdouble')
