+++ quaddtype/numpy_quaddtype/src/scalar.c
@@ -1,0 +1,159 @@
+#include <Python.h>
+#include <sleef.h>
+#include <sleefquad.h>
+#include <stdlib.h>
+
+#define PY_ARRAY_UNIQUE_SYMBOL QuadPrecType_ARRAY_API
+#define NPY_NO_DEPRECATED_API NPY_2_0_API_VERSION
+#define NO_IMPORT_ARRAY
+
+#include "numpy/arrayobject.h"
+#include "numpy/ndarraytypes.h"
+#include "numpy/dtype_api.h"
+
+#include "scalar.h"
+// #include "scalar_ops.h"
+
+QuadPrecisionObject *
+QuadPrecision_raw_new(QuadBackendType backend)
+{
+    QuadPrecisionObject *new = PyObject_New(QuadPrecisionObject, &QuadPrecision_Type);
+    if (!new)
+        return NULL;
+    new->backend = backend;
+    if (backend == BACKEND_SLEEF) {
+        new->value.sleef_value = Sleef_cast_from_doubleq1(0.0);
+    }
+    else {
+        new->value.longdouble_value = 0.0L;
+    }
+    return new;
+}
+
+QuadPrecisionObject *
+QuadPrecision_from_object(PyObject *value, QuadBackendType backend)
+{
+    QuadPrecisionObject *self = QuadPrecision_raw_new(backend);
+    if (!self)
+        return NULL;
+
+    if (PyFloat_Check(value)) {
+        double dval = PyFloat_AsDouble(value);
+        if (backend == BACKEND_SLEEF) {
+            self->value.sleef_value = Sleef_cast_from_doubleq1(dval);
+        }
+        else {
+            self->value.longdouble_value = (long double)dval;
+        }
+    }
+    else if (PyUnicode_CheckExact(value)) {
+        const char *s = PyUnicode_AsUTF8(value);
+        char *endptr = NULL;
+        if (backend == BACKEND_SLEEF) {
+            self->value.sleef_value = Sleef_strtoq(s, &endptr);
+        }
+        else {
+            self->value.longdouble_value = strtold(s, &endptr);
+        }
+        if (*endptr != '\0' || endptr == s) {
+            PyErr_SetString(PyExc_ValueError, "Unable to parse string to QuadPrecision");
+            Py_DECREF(self);
+            return NULL;
+        }
+    }
+    else if (PyLong_Check(value)) {
+        long long val = PyLong_AsLongLong(value);
+        if (val == -1 && PyErr_Occurred()) {
+            PyErr_SetString(PyExc_OverflowError, "Overflow Error, value out of range");
+            Py_DECREF(self);
+            return NULL;
+        }
+        if (backend == BACKEND_SLEEF) {
+            self->value.sleef_value = Sleef_cast_from_int64q1(val);
+        }
+        else {
+            self->value.longdouble_value = (long double)val;
+        }
+    }
+    else {
+        PyErr_SetString(PyExc_TypeError, "QuadPrecision value must be a float, int or string");
+        Py_DECREF(self);
+        return NULL;
+    }
+
+    return self;
+}
+
+static PyObject *
+QuadPrecision_new(PyTypeObject *cls, PyObject *args, PyObject *kwargs)
+{
+    PyObject *value;
+    const char *backend_str = "sleef";
+    static char *kwlist[] = {"value", "backend", NULL};
+
+    if (!PyArg_ParseTupleAndKeywords(args, kwargs, "O|s", kwlist, &value, &backend_str)) {
+        return NULL;
+    }
+
+    QuadBackendType backend = BACKEND_SLEEF;
+    if (strcmp(backend_str, "longdouble") == 0) {
+        backend = BACKEND_LONGDOUBLE;
+    }
+    else if (strcmp(backend_str, "sleef") != 0) {
+        PyErr_SetString(PyExc_ValueError, "Invalid backend. Use 'sleef' or 'longdouble'.");
+        return NULL;
+    }
+
+    return (PyObject *)QuadPrecision_from_object(value, backend);
+}
+
+static PyObject *
+QuadPrecision_str(QuadPrecisionObject *self)
+{
+    char buffer[128];
+    if (self->backend == BACKEND_SLEEF) {
+        Sleef_snprintf(buffer, sizeof(buffer), "%.*Qe", SLEEF_QUAD_DIG, self->value.sleef_value);
+    }
+    else {
+        snprintf(buffer, sizeof(buffer), "%.35Le", self->value.longdouble_value);
+    }
+    return PyUnicode_FromString(buffer);
+}
+
+static PyObject *
+QuadPrecision_repr(QuadPrecisionObject *self)
+{
+    PyObject *str = QuadPrecision_str(self);
+    if (str == NULL) {
+        return NULL;
+    }
+    const char *backend_str = (self->backend == BACKEND_SLEEF) ? "sleef" : "longdouble";
+    PyObject *res = PyUnicode_FromFormat("QuadPrecision('%S', backend='%s')", str, backend_str);
+    Py_DECREF(str);
+    return res;
+}
+
+static void
+QuadPrecision_dealloc(QuadPrecisionObject *self)
+{
+    Py_TYPE(self)->tp_free((PyObject *)self);
+}
+
+PyTypeObject QuadPrecision_Type = {
+        PyVarObject_HEAD_INIT(NULL, 0).tp_name = "numpy_quaddtype.QuadPrecision",
+        .tp_basicsize = sizeof(QuadPrecisionObject),
+        .tp_itemsize = 0,
+        .tp_new = QuadPrecision_new,
+        .tp_dealloc = (destructor)QuadPrecision_dealloc,
+        .tp_repr = (reprfunc)QuadPrecision_repr,
+        .tp_str = (reprfunc)QuadPrecision_str,
+        // .tp_as_number = &quad_as_scalar,
+        // .tp_richcompare = (richcmpfunc)quad_richcompare
+
+};
+
+int
+init_quadprecision_scalar(void)
+{
+    return PyType_Ready(&QuadPrecision_Type);
+}
