+++ quaddtype/numpy_quaddtype/src/scalar_ops.cpp
