+++ quaddtype/numpy_quaddtype/src/scalar_ops.h
