+++ quaddtype/numpy_quaddtype/src/umath.cpp
